Ya casi se puede crear el pago

+++ PagoAgil/Aplicacion/Modelo/ClienteSQL/PagoDB.cs
@@ -9,13 +9,13 @@ namespace PagoAgil.Aplicacion.Modelo.ClienteSQL
     public class PagoDB
     {
         public long numeroPago { get; set; }
-        public double montoTotal { get; set; }
+        public decimal montoTotal { get; set; }
         public DateTime fechaPago { get; set; }
         public int idMedioPago { get; set; }
         public int dniCliente { get; set; }
         public int cpSucursal { get; set; }
 
-        public PagoDB(long unNroPago, double unMonto, DateTime unaFechaPago, int unIdMedioPago, int unDniCliente, int unCpSucursal)
+        public PagoDB(long unNroPago, decimal unMonto, DateTime unaFechaPago, int unIdMedioPago, int unDniCliente, int unCpSucursal)
         {
             this.numeroPago = unNroPago;
             this.montoTotal = unMonto;
